python-seeder/requirements.txt
ijson==3.2.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
tqdm==4.66.1